requirements.txt
django~=4.0.2
gunicorn
django-heroku
pillow~=9.0.1
django-multiselectfield
requests~=2.27.1
spoonacular